+++ www/apps/api-reference/specs/store/code_samples/Shell/store_store-credit-accounts_claim/post.sh
@@ -1,0 +1,7 @@
+curl -X POST '{backend_url}/store/store-credit-accounts/claim' \
+-H 'Authorization: Bearer {access_token}' \
+-H 'x-publishable-api-key: {your_publishable_api_key}' \
+-H 'Content-Type: application/json' \
+--data-raw '{
+  "code": "{value}"
+}'