+++ www/packages/docs-ui/src/global-config.ts
@@ -4,8 +4,8 @@ import { DocsConfig } from "types"
 
 export const globalConfig: Pick<DocsConfig, "version"> = {
   "version": {
-    "number": "2.13.6",
-    "releaseUrl": "https://github.com/medusajs/medusa/releases/tag/v2.13.6",
-    "releaseDate": "2026-04-08T09:01:38Z"
+    "number": "2.14.1",
+    "releaseUrl": "https://github.com/medusajs/medusa/releases/tag/v2.14.1",
+    "releaseDate": "2026-04-24T14:11:43Z"
   }
 }