+++ packages/admin/admin-vite-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/admin-vite-plugin
 
+## 2.14.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/admin-shared@2.14.1
+
+## 2.14.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/admin-shared@2.14.0
+
 ## 2.13.6
 
 ### Patch Changes